+++ docs/theme.config.jsx
@@ -1,7 +1,7 @@
 export default {
   logo: <span>UnstoppableSwap</span>,
   project: {
-    link: "https://github.com/UnstoppableSwap/unstoppableswap-docs",
+    link: "https://github.com/UnstoppableSwap/core",
   },
   primaryHue: 14.3,
   primarySaturation: 90.68,